requirements.txt
asgiref==3.8.1
Django==5.1.4
djangorestframework==3.15.2
sqlparse==0.5.3
typing_extensions==4.12.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.10
sentry-sdk==2.19.0